moved tests to be consistent w/ organization of tests_tf folder
